# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "后台管理系统",
        "title": "backend-admin",
        "contact": {},
        "version": "v1"
    },
    "paths": {
        "/v1/sys/api/:id": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新api接口",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "更新api接口",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SysApi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/api/generate": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "生成接口",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "生成接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/api/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "接口列表",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "接口列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.SysApi"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/sys/menu": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "新增菜单",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "新增菜单",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddMenuParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/menu/:id": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新菜单",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMenuParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除菜单",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "删除菜单",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/menu/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "菜单列表",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "菜单列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页面",
                        "name": "page_no",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "description": "页大小",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RoleListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/sys/role": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "增加角色",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "增加角色",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddRoleParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/role/:id": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新角色",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "更新角色",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateRoleParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除角色",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "删除角色",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/v1/sys/role/list": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "角色列表",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "角色列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页面",
                        "name": "page_no",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 200,
                        "minimum": 1,
                        "type": "integer",
                        "description": "页大小",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.RoleListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddMenuParam": {
            "type": "object",
            "required": [
                "menu_name",
                "menu_type",
                "status"
            ],
            "properties": {
                "apis": {
                    "description": "api接口id集合",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "is_show": {
                    "description": "是否显示1:正常 0:停用",
                    "type": "integer"
                },
                "menu_name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "menu_type": {
                    "description": "菜单类型 1:目录 2:菜单 3:按钮",
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                },
                "parent_id": {
                    "description": "父级菜单",
                    "type": "integer"
                },
                "path": {
                    "description": "菜单路径",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer",
                    "minimum": 0
                },
                "status": {
                    "description": "状态 -1:删除 1:正常 2:停用",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "title": {
                    "description": "菜单标题",
                    "type": "string"
                }
            }
        },
        "dto.AddRoleParam": {
            "type": "object",
            "required": [
                "menu_ids",
                "role_key",
                "role_name"
            ],
            "properties": {
                "flag": {
                    "description": "标志位",
                    "type": "string"
                },
                "menu_ids": {
                    "description": "菜单id集合",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "role_key": {
                    "description": "角色唯一key",
                    "type": "string"
                },
                "role_name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "role_sort": {
                    "description": "角色排序",
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "dto.RoleListResp": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "数据总数",
                    "type": "integer"
                },
                "data": {
                    "description": "数据集合",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SysRole"
                    }
                }
            }
        },
...
        "dto.UpdateMenuParam": {
            "type": "object",
            "required": [
                "menu_name",
                "menu_type",
                "status"
            ],
            "properties": {
                "apis": {
                    "description": "api接口id集合",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "is_show": {
                    "description": "是否显示1:正常 0:停用",
                    "type": "integer"
                },
                "menuId": {
                    "description": "id",
                    "type": "integer"
                },
                "menu_name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "menu_type": {
                    "description": "菜单类型 1:目录 2:菜单 3:按钮",
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                },
                "parent_id": {
                    "description": "父级菜单",
                    "type": "integer",
                    "minimum": 0
                },
                "path": {
                    "description": "菜单路径",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer",
                    "minimum": 0
                },
                "status": {
                    "description": "状态 -1:删除 1:正常 2:停用",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "title": {
                    "description": "菜单标题",
                    "type": "string"
                }
            }
        },
        "dto.UpdateRoleParam": {
            "type": "object",
            "required": [
                "menu_ids",
                "role_name"
            ],
            "properties": {
                "flag": {
                    "description": "标志位",
                    "type": "string"
                },
                "menu_ids": {
                    "description": "菜单id集合",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "role_id": {
                    "description": "角色编码",
                    "type": "integer"
                },
                "role_name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "role_sort": {
                    "description": "RoleKey  *string `form:\"role_key\" json:\"role_key\"  validate:\"required\" `  //角色key不能修改",
                    "type": "integer"
                }
            }
        },
...
        "models.SysApi": {
            "type": "object",
            "properties": {
                "createBy": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "handle": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updateBy": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.SysRole": {
            "type": "object",
            "properties": {
                "admin": {
                    "description": "是否是管理员角色",
                    "type": "boolean"
                },
                "createBy": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "flag": {
                    "description": "标志位",
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "roleId": {
                    "description": "角色编码",
                    "type": "integer"
                },
                "roleKey": {
                    "description": "角色唯一key 不能修改",
                    "type": "string"
                },
                "roleName": {
                    "description": "角色名称",
                    "type": "string"
                },
                "roleSort": {
                    "description": "角色排序",
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updateBy": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Detail string `json:\"detail\"`"
                },
                "msg": {
                    "type": "string"
                },
                "ret": {
                    "type": "integer"
                }
            }
        }
    }
}